+++ Moloco/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.3.0.0
+* Certified with Moloco SDK 3.3.0.
+* Annotated all `Activity` parameters with `@Nullable` to better avoid potential NPEs.
+
 ## 3.2.0.0
 * Certified with Moloco SDK 3.2.0.